# Optimize SHA3 implementation.