--- a/src/common.jl
+++ b/src/common.jl
@@ -13,9 +13,7 @@ function update!(context::T, data::U) where {T<:Union{SHA1_CTX,SHA2_CTX,SHA3_CTX
     usedspace = context.bytecount % blocklen(T)
     while len - data_idx + usedspace >= blocklen(T)
         # Fill up as much of the buffer as we can with the data given us
-        for i in 1:(blocklen(T) - usedspace)
-            context.buffer[usedspace + i] = data[data_idx + i]
-        end
+        copy!(context.buffer, usedspace + 1, data, data_idx + 1, blocklen(T) - usedspace)
 
         transform!(context)
         context.bytecount += blocklen(T) - usedspace
@@ -25,9 +23,7 @@ function update!(context::T, data::U) where {T<:Union{SHA1_CTX,SHA2_CTX,SHA3_CTX
 
     # There is less than a complete block left, but we need to save the leftovers into context.buffer:
     if len > data_idx
-        for i = 1:(len - data_idx)
-            context.buffer[usedspace + i] = data[data_idx + i]
-        end
+        copy!(context.buffer, usedspace + 1, data, data_idx + 1, len - data_idx)
         context.bytecount += len - data_idx
     end
 end