This makes SHA3 about 3x faster, and the other hash functions moderately faster. The code is a little uglier, but worth it I think!

Hashing a ~400MB file:

**Before**
```
read:      0.169224 seconds (6 allocations: 392.499 MiB, 26.97% gc time)
SHA-1:     1.788647 seconds (7 allocations: 960 bytes)
SHA2-256:   4.946575 seconds (6 allocations: 512 bytes)
SHA2-512:   3.785016 seconds (6 allocations: 656 bytes)
SHA3-256:  15.608980 seconds (3.03 M allocations: 369.411 MiB, 0.02% gc time)
SHA3-512:  29.399319 seconds (5.72 M allocations: 697.775 MiB, 0.02% gc time)
```

**After**
```
read:      0.171445 seconds (6 allocations: 392.499 MiB, 26.98% gc time)
SHA-1:     1.358481 seconds (7 allocations: 960 bytes)
SHA2-256:   4.567648 seconds (6 allocations: 512 bytes)
SHA2-512:   2.906533 seconds (6 allocations: 656 bytes)
SHA3-256:   5.006596 seconds (8 allocations: 992 bytes)
SHA3-512:   9.348853 seconds (8 allocations: 960 bytes)
```

cc: @staticfloat