@staticfloat 2018-01-22
Thanks for this @dcjones!  Can you post your timing script here for posterity's sake?