@dcjones 2018-01-22
It's actually your timing script in `test/perf.jl`. The (arbitrary) file I was using to test is this one, after decompressing: [ftp://ftp.ensembl.org/pub/release-91/gff3/homo_sapiens/Homo_sapiens.GRCh38.91.gff3.gz]()